refactor: improve naming #97
