+++ src/features/Search/components/VisualNovelList/VisualNovelList.tsx
@@ -3,8 +3,8 @@ import { Box } from '@chakra-ui/react';
 import { VisualNovel } from '../../../../models/visualNovels/visualNovel';
 import { CoverCardSkeleton } from '../CoverCard/CoverCardSkeleton';
 import { CoverCard } from '../CoverCard/CoverCard';
-import { VNTable } from '../VNTable/VNTable';
-import { VNTableSkeleton } from '../VNTable/VNTableSkeleton';
+import { VNTable } from '../VisualNovelTable/VNTable';
+import { VNTableSkeleton } from '../VisualNovelTable/VisualNovelTableSkeleton';
 import { WideCard } from '../WideCard/WideCard';
 import { WideCardSkeleton } from '../WideCard/WideCardSkeleton';
 import { ExtendedCard } from '../ExtendedCard/ExtendedCard';
