+++ src/utils/randomWidth.ts
@@ -1,6 +1,6 @@
 /**
- * Get random number.
+ * Gets random width in percent.
  * @param min Minimum random value.
  * @param max Maximum random number.
  */
-export const randomWidth = (min = 25, max = 100): string => `${Math.round((Math.random() * (max - min)) + min)}%`;
+export const getRandomWidth = (min = 25, max = 100): string => `${Math.round((Math.random() * (max - min)) + min)}%`;
